Refactor(Processing): Remove superflous interfaces
